+++ lyc_flutter_project/lib/common/widget/content_box.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+
+class ContentBox extends StatelessWidget {
+  final String title;
+  final Widget child;
+
+  const ContentBox({
+    super.key,
+    required this.title,
+    required this.child,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(
+        bottom: 20.0,
+      ),
+      padding: const EdgeInsets.symmetric(
+        vertical: 16.0,
+        horizontal: 20.0,
+      ),
+
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(
+          20,
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(8),
+            child: Text(
+              title,
+              style: const TextStyle(
+                fontWeight: FontWeight.w600,
+                fontSize: 16,
+              ),
+            ),
+          ),
+          child,
+        ],
+      ),
+    );
+  }
+}
